ADD: update examplechip to show changes proposed
