+++ common/src/hal_clint.c
@@ -10,20 +10,20 @@
 
 #include "hal_clint.h"
 
-uint64_t HAL_CLINT_getTime() {
+uint64_t HAL_CLINT_getTime(CLINT_TypeDef *CLINTx) {
   uint32_t time_lo;
   uint32_t time_hi;
 
   do {
-    time_hi = *((volatile uint32_t *)(&CLINT->MTIME) + 1);
-    time_lo = *((volatile uint32_t *)(&CLINT->MTIME));
-  } while (*((volatile uint32_t *)(&CLINT->MTIME) + 1) != time_hi);
+    time_hi = *((volatile uint32_t *)(CLINTx->MTIME) + 1);
+    time_lo = *((volatile uint32_t *)(CLINTx->MTIME));
+  } while (*((volatile uint32_t *)(CLINTx->MTIME) + 1) != time_hi);
 
   return (((uint64_t)time_hi) << 32U) | time_lo;
 }
 
-void HAL_CLINT_setTimerInterruptTarget(uint32_t hartid, uint64_t time) {
-  *((volatile uint32_t *)(&CLINT->MTIMECMP0) + 4 * hartid + 1) = 0xffffffff;
-  *((volatile uint32_t *)(&CLINT->MTIMECMP0) + 4 * hartid) = (uint32_t)time;
-  *((volatile uint32_t *)(&CLINT->MTIMECMP0) + 4 * hartid + 1) = (uint32_t)(time >> 32);
+void HAL_CLINT_setTimerInterruptTarget(CLINT_TypeDef *CLINTx, uint32_t hartid, uint64_t time) {
+  *((volatile uint32_t *)(CLINTx->MTIMECMP0) + 4 * hartid + 1) = 0xffffffff;
+  *((volatile uint32_t *)(CLINTx->MTIMECMP0) + 4 * hartid) = (uint32_t)time;
+  *((volatile uint32_t *)(CLINTx->MTIMECMP0) + 4 * hartid + 1) = (uint32_t)(time >> 32);
 }
