--- examplechip/src/examplechip_hal.c
@@ -1,17 +1,0 @@
-/**
- * @file hal.c
- * @author -T.K.- / t_k_233@outlook.com
- * @brief 
- * @version 0.1
- * 
- * @copyright Copyright (c) 2023
- * 
- */
-
-#include "examplechip_hal.h"
-
-void HAL_init() {
-  // for (uint16_t i=0; i<64; i+=1) {
-  //   HAL_PLIC_completeIRQ(0, i);
-  // }
-}
