+++ examplechip/src/hal_rcc.c
@@ -8,6 +8,6 @@
  * 
  */
 
-#include "examplechip_hal_rcc.h"
+#include "hal_rcc.h"
 
 void HAL_RCC_InitSystemClock() {}
